Some more PR comment fixes
