+++ src/lib.rs
@@ -507,10 +507,7 @@ pub mod prelude {
             channels::{ChannelKind, RepliconChannel, RepliconChannels},
             command_markers::AppMarkerExt,
             common_conditions::*,
-            connected_clients::{
-                client_visibility::ClientVisibility, ConnectedClients, ReplicatedClient,
-                ReplicatedClients, VisibilityPolicy,
-            },
+            connected_clients::ConnectedClients,
             event_registry::{
                 client_event::{ClientEventAppExt, FromClient},
                 server_event::{SendMode, ServerEventAppExt, ToClients},
@@ -530,6 +527,10 @@ pub mod prelude {
     pub use super::server::{
         client_entity_map::{ClientEntityMap, ClientMapping},
         events::ServerEventsPlugin,
+        replicated_clients::{
+            client_visibility::ClientVisibility, ReplicatedClient, ReplicatedClients,
+            VisibilityPolicy,
+        },
         ServerEvent, ServerPlugin, ServerSet, StartReplication, TickPolicy,
     };
